% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Firstfonction.R
\name{Addition}
\alias{Addition}
\alias{Multiplication}
\title{Calculs basiques}
\usage{
Addition(x, y)

Multiplication(x, y)
}
\arguments{
\item{x}{un nombre}

\item{y}{un nombre}
}
\value{
la somme ou le produit de \code{x} et \code{y}
}
\description{
Calculs basiques
}
\examples{
Addition(4,7)
}
