% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEtuto-package.R
\docType{package}
\name{MEtuto-package}
\alias{MEtuto}
\alias{MEtuto-package}
\title{MEtuto: Exemples for package creation}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Show exemples for the creation of a package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasvincent79/MEtuto}
  \item Report bugs at \url{https://github.com/thomasvincent79/MEtuto/issues}
}

}
\author{
\strong{Maintainer}: Me MyFamily \email{Met@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
